Current progress: db created with failing tests
